requirement.txt
easyocr
opencv-python
numpy
matplotlib
pillow
google-genai
python-dotenv
